requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
nltk==3.8.1
numpy<2.0
scikit-learn==1.3.0
pickle-mixin==1.0.2
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
torch==2.1.2
torchvision==0.16.2
typing-extensions<4.6.0,>=4.5.0
transformers==4.35.2
waitress